Prevent LTFO from deleting itself while updating

+++ LTFO.py
@@ -73,7 +73,8 @@ def update():
     try: #remove previous version if just updated
         global proj
         with open(proj+'.tmp', 'r') as content_file:
-            os.remove(str(content_file.read()))
+            if content_file.read() != os.path.basename(__file__): #if the old version has the current filename, don't delete
+                os.remove(str(content_file.read()))
         os.remove(proj+'.tmp')
     except:
         pass
